+++ chatinterface_server/internal/config.py
@@ -56,66 +56,6 @@ def load_or_create_config(config_path: str, default_config: dict) -> dict:
             return json.load(file)
 
 
-class ConfigMaker:
-    def __init__(self, log_dir: str, conf_dir: str) -> None:
-        self.log_dir: str = log_dir
-        self.conf_dir: str = conf_dir
-
-    def create_logging_config(self, log_level: str, use_console: bool = True) -> dict:
-        handlers: dict = {
-            'console': {
-                'class': 'logging.StreamHandler',
-                'stream': 'ext://sys.stdout',
-                'formatter': 'default'
-            }
-        }
-        loggers: dict = {}
-
-        log_targets: set[str] = {
-            'ws', 'auth', 'chats', 
-            'database', 'main'
-        }
-        for target_name in log_targets:
-            logfile: str = os.path.join(self.log_dir, f'{target_name}.log')
-            handler_name: str = f'chatinterface.handler.{target_name}'
-
-            logger_name: str = f'chatinterface.logger.{target_name}'
-            handler_list: list[str] = [handler_name]
-
-            if use_console:
-                handler_list.append('console')
-
-            handlers[handler_name] = {
-                'class': 'logging.FileHandler',
-                'formatter': 'default',
-                'filename': logfile,
-            }
-            loggers[logger_name] = {
-                'handlers': handler_list,
-                'level': log_level,
-                'propagate': False
-            }
-
-        return {
-            'version': 1,
-            'formatters': {
-                'default': {
-                    'format': '[%(name)s]: [%(funcName)s] - [%(asctime)s] - [%(levelname)s] - %(message)s',
-                    'datefmt': '%Y-%m-%d %H:%M:%S'
-                },
-                'precise': {
-                    'format': (
-                        '[%(name)s, %(funcName)s] - [%(levelname)s] - "%(pathname)s:%(lineno)d"'
-                        ' - [%(asctime)s]: %(message)s'
-                    ),
-                    'datefmt': '%Y-%m-%d %H:%M:%S'
-                }
-            },
-            'handlers': handlers,
-            'loggers': loggers
-        }
-
-
 class ConfigManager:
     def __init__(self, base_dir: str = DEFAULT_APP_DIR) -> None:
         if not base_dir:
@@ -125,26 +65,56 @@ class ConfigManager:
 
         # if two instances are being ran, version separates instances if base dir is same
         abs_base_dir: str = os.path.abspath(base_dir)
-        ver_base_dir: str = os.path.join(abs_base_dir, __version__)
+        self.base_dir: str = os.path.join(abs_base_dir, __version__)
 
-        log_dir: str = os.path.join(ver_base_dir, 'logs')
-        conf_dir: str = os.path.join(ver_base_dir, 'config')
+        os.makedirs(self.base_dir, mode=0o700, exist_ok=True)
 
-        os.makedirs(ver_base_dir, mode=0o700, exist_ok=True)
-        os.makedirs(log_dir, mode=0o700, exist_ok=True)
+    def make_logging_config(self):
+        return {
+            'version': 1,
+            'formatters': {
+                'default': {
+                    'format': '[%(name)s]: [%(module)s | %(funcName)s] - [%(asctime)s] - [%(levelname)s] - %(message)s',
+                    'datefmt': '%Y-%m-%d %H:%M:%S'
+                },
+                'precise': {
+                    'format': (
+                        '[%(name)s]: [%(module)s | %(funcName)s] - [%(levelname)s] - "%(pathname)s:%(lineno)d"'
+                        ' - [%(asctime)s]: %(message)s'
+                    ),
+                    'datefmt': '%Y-%m-%d %H:%M:%S'
+                }
+            },
+            'handlers': {
+                'console': {
+                    'class': 'logging.StreamHandler',
+                    'stream': 'ext://sys.stdout',
+                    'formatter': 'default'
+                },
+                "chatinterface_server": {
+                    "class": "logging.handlers.RotatingFileHandler",
+                    "formatter": "default",
+                    "maxBytes": 5 * 1024 * 1024,  # 5 MB
+                    "backupCount": 3,
+                    "filename": os.path.join(self.base_dir, 'chatinterface_server.log')
+                }
+            },
+            'loggers': {
+                "chatinterface_server": {
+                    "handlers": [
+                        "chatinterface_server",
+                        "console"
+                    ],
+                    "level": "INFO",
+                    "propagate": True
+                }
+            }
+        }
 
-        os.makedirs(conf_dir, mode=0o700, exist_ok=True)
+    def setup_logging(self) -> None:
+        log_config_file: str = os.path.join(self.base_dir, 'logging.json')
 
-        self.log_dir: str = log_dir
-        self.conf_dir: str = conf_dir
-
-        self.conf_maker: ConfigMaker = ConfigMaker(self.log_dir, self.conf_dir)
-
-    def setup_logging(self, log_level: str = "INFO") -> None:
-        log_config_file: str = os.path.join(self.conf_dir, 'log.json')
-        default_logging_config: dict = self.conf_maker.create_logging_config(log_level)
-
-        log_config: dict = load_or_create_config(log_config_file, default_logging_config)
+        log_config: dict = load_or_create_config(log_config_file, self.make_logging_config())
         logging.config.dictConfig(log_config)