+++ chatinterface_server/models/common.py
@@ -1,13 +1,15 @@
 import typing
-from typing import NamedTuple
+from typing import Annotated, NamedTuple
 from fastapi.templating import Jinja2Templates
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 if typing.TYPE_CHECKING:
     from ..internal.database import MainDatabase
     from ..internal.config import ConfigManager
     from ..internal.ws import WebsocketClients
 
+
+UsernameField = Annotated[str, Field(max_length=20, min_length=1)]
 
 class SessionInfo(BaseModel):
     username: str