+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Implement message composing
+# Create /api/users for user management
 
 **Version**: v0.1.0
 
@@ -6,9 +6,29 @@
 
 ## Additions
 
-**`internal/constants.py`**:
+**`/main.py`**:
 
-* Create `MESSAGE_COMPOSE` constant.
+* Added the users router to the main app.
+
+**`internal/database.py`**:
+
+* Created `delete_user` and `get_users` in `UserMethods` for user management.
+
+**`internal/ws.py`**:
+
+* Created `disconnect_all_clients` which disconnects all clients by username.
+
+**`models/users.py`**:
+
+* Created for the `/users` router.
+
+**`routers/users.py`**:
+
+* Created `/users` endpoint for user management.
+
+**`docker/docker-compoose.yml`**:
+
+* Created volume mount which links `./chatinterface_server` to `/app/chatinterface-server_config`.
 
 ## Changes
 
@@ -16,7 +36,23 @@
 
 * Changed `/compose_message` to mimic `/send_message` but for composing.
 
+**`internal/config.py`**:
+
+* Rewrote the logging setup, now only uses `chatinterface_logger` and rewrote `ConfigManager` to only handle the logging configuration.
+
+**`internal/database.py`**:
+
+* Changed `UserMethods.add_user` to return True when successful.
+
+**`models/ws.py`**:
+
+* Moved `UsernameField` to `models/common.py`.
+
+**`docker/Dockerfile`**:
+
+* Changed entrypoint from `uvicorn chatinterface_server.main:app` to `fastapi run chatinterface_server/main.py`.
+
 ## Misc
 
-* Frontend JavaScript updated to support message composing.
+* Frontend JavaScript now shows a message if you are logged out.
 * Will create a user management page soon.